refactor: extract adapter interfaces into subpackage and assert them
The optional interfaces that can be implemented by an adapter are now
located in the adapter/port package. This allows the adapter
implementations to import and assert that those interfaces are
implemented.
