+++ internal/adapter/outlook_http/adapter.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pkg/browser"
 	"golang.org/x/oauth2"
 
+	"github.com/inovex/CalendarSync/internal/adapter/port"
 	"github.com/inovex/CalendarSync/internal/auth"
 	"github.com/inovex/CalendarSync/internal/models"
 )
@@ -40,6 +41,11 @@ type CalendarAPI struct {
 
 	storage auth.Storage
 }
+
+// Assert that the expected interfaces are implemented
+var _ port.Configurable = &CalendarAPI{}
+var _ port.LogSetter = &CalendarAPI{}
+var _ port.OAuth2Adapter = &CalendarAPI{}
 
 func (c *CalendarAPI) SetupOauth2(ctx context.Context, credentials auth.Credentials, storage auth.Storage, bindPort uint) error {
 	// Outlook Adapter does not need the clientKey
